requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
feedparser==6.0.2
html2text==2020.1.16
idna==2.10
importlib-resources==3.3.0
infi.systray==0.1.12
Pillow==8.0.1
playsound==1.2.2
pystray==0.17.1
realpython-reader==1.0.0
requests==2.25.0
schedule==0.6.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
typing==3.7.4.3
urllib3==1.26.2
